requirements.txt
playwright>=1.40.0
pytest>=7.4.0
python-dotenv>=1.0.0
argparse
tenacity>=8.2.3
